backend/requirements.txt
flask
flask-cors
Pillow
qrcode
opencv-python
PyQt5
